+++ example/android/app/src/main/kotlin/com/famproperties/amazon_s3_cognito_example/MainActivity.kt
@@ -6,8 +6,5 @@ import io.flutter.app.FlutterActivity
 import io.flutter.plugins.GeneratedPluginRegistrant
 
 class MainActivity: FlutterActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    GeneratedPluginRegistrant.registerWith(this)
-  }
+
 }