+++ CHANGELOG.md
@@ -38,3 +38,20 @@ the documents and image urls path are now returned correctly.
 ## 0.6.0
 prerelease -> this version includes null safety
 
+## 0.7.0
+Breaking changes area added.
+1) fixed issue : The plugins amazon_s3_cognito, version of the Android embedding
+2) Added methods that can upload multiple images via a list
+3) When using multiple image upload, we can listen to image upload using event stream
+4) for IOS app now uses TransferUtility API's, since TransferManager API's are depracated
+5) Now all methods take subFolder to which we want to upload/ delete image.
+
+## 0.7.3
+1) fixed issue that on IOS image upload was not working.
+
+## 0.7.4
+1) fixed bug - In case image upload failed due to file not found exception on android,
+then image upload did not succeed and app was not returning failure back incase of multiple image upload.
+
+## 0.7.5
+1) Plugin is compatible now with flutter 3.0 and above.